+++ networkmodel/network.py
@@ -3,10 +3,10 @@
 import numpy as np
 import pandas as pd
 
-from networkmodel.buildmat import site_key
+from networkmodel.BuildMatrix import site_key
 from networkmodel.config import TIME_POINTS_PROTEIN, MODEL, RESULTS_DIR
 from networkmodel.models import distributive_rhs, build_random_transitions, sequential_rhs, combinatorial_rhs
-from networkmodel.steadystate import build_y0_from_data
+from networkmodel.SteadyState import build_y0_from_data
 from config.config import setup_logger
 
 logger = setup_logger(log_dir=RESULTS_DIR)
